Use EXCLUDE_FROM_ALL when adding subdirectories
It seems that Zephyr tries to build `all` target no matter what we actually depend on.
But e.g. some targets in common require that specific peripheral is defined.
With  EXCLUDE_FROM_ALL only explicit dependencies are linked to the final image.
So we can remove those `if(CONFIG_ADC) ...` in common/CMakeLists.txt.